translation_microservice/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
googletrans==4.0.0rc1
python-dotenv==1.0.0
aiosqlite==0.19.0
pytest==7.4.3
httpx==0.25.2